# perf: Optimize hash calculation with buffer implementation